requirements.txt
anthropic==0.50.0
beautifulsoup4==4.13.4
gradio_client==1.10.0
json5==0.12.0
numpy==2.2.5
openai==1.76.2
opencv_python==4.11.0.86
Pillow==11.2.1
protobuf==6.30.2
regex==2024.11.6
Requests==2.32.3
rouge==1.0.1
selenium==4.31.0
tqdm==4.66.1
webdriver_manager==4.0.2
zhipuai==2.1.5.20250421
